==> scripts/04c_models_proportions_inference_2022.Rmd <==
---
title: "Activity center models - days with detections (inference) - 2022"
author: "Cara"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true
    toc_float: true
    toc_depth: 4
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r load-packages, include=FALSE}
library(data.table)
library(effects)
library(ggplot2)
library(lme4)
# 
# library(plyr)
# library(broom)
# library(tidyverse)
# library(kableExtra)
# library(lme4)
# library(glmmTMB)
# library(ggeffects)
```

------------------------------------------------------------------------

### Preliminary models for analyzing spotted owl activity center ("big grid") data from 2021

------------------------------------------------------------------------

#### Load data:

##### 145 rows = 145 ARU stations from 4 different study areas

```{r echo=FALSE}
  #import nDays summary
  daysSTOC <- fread('../output/09_ac_22_nights_with_det.csv')

  #import station summary
  acData <- fread('../output/05_ac_2022_merged.csv')
  
  #merge some columns and order level
  daysMerge <- merge(daysSTOC[,-c('V1')], acData[,c('SITE_STN','reproState','Distance','dist_intended')])
  daysMerge$reproState <- factor(daysMerge$reproState, levels = c('female','pair'))

  #add activity center ID field
  daysMerge$siteID <- sapply(strsplit(as.character(daysMerge$SITE_STN), '\\_'), '[', 1)
  daysMerge$siteID <- factor(daysMerge$siteID, levels = c('CC','DC','BC','LM'))
  
  #add nickname for presentations
  daysMerge$nname <- ifelse(daysMerge$siteID %in% 'DC', 'P1',
                            ifelse(daysMerge$siteID %in% 'BC', 'P2',
                                   ifelse(daysMerge$siteID %in% 'LM','P3','Fe1')))
  daysMerge$nname <- factor(daysMerge$nname, levels = c('Fe1','P1','P2','P3'))
  
  daysMerge[,-c('dist_intended')]
    
```

------------------------------------------------------------------------

### -- Models using the proportion of survey nights with any spotted owl detection as the response variable

#### 1. Quasibinomial with distance from center as covariate

```{r}

mod1 <- glm(cbind(nights_STOC_ANY, duration-nights_STOC_ANY) ~ Distance,
            data = daysMerge, family = quasibinomial(link = 'logit'))
summary(mod1)

```

plot:

```{r echo=FALSE}
plot.mod1b <- ggplot(daysMerge, aes(x = Distance/1000, y = nights_STOC_ANY/duration,
                      nights = nights_STOC_ANY, total = duration)) +
       geom_point(data = daysMerge[daysMerge$nights_STOC_ANY > 0,], #we don't want to plot 0s
                  aes(x = Distance/1000, y = nights_STOC_ANY / duration),
                  color = 'gray40', size = 2) +
       geom_smooth(formula = cbind(nights,total-nights) ~ x, 
                   method = 'glm', method.args = list(family = quasibinomial(link = 'logit')),
                   color = 'aquamarine4', fill = 'aquamarine3') +
              ylim(0,0.65) +
       # ggtitle('Quasibinomial') + 
       ylab('Proportion of nights with detection') + xlab('Distance from center (km)') +
       # theme_bw() +
  theme(panel.background = element_rect(fill = 'transparent'),
          panel.grid = element_line(color = 'gray90'),
          axis.line = element_line(),
          axis.title = element_text(size = 18),
          # axis.title.x = element_blank(),
          strip.text = element_text(size = 16),
          axis.text = element_text(size = 16),
          plot.title = element_text(hjust = 0.5),
          legend.text = element_text(size = 16),
          legend.title = element_blank(),
          legend.position = 'none',
          legend.background = element_rect(fill='transparent'))
plot.mod1
# plot(allEffects(mod1))

```


#### 2. Add a fixed effect for AC site
##### (CC = Cummins Creek, DC = Drift Creek, BC = Baker Creek, LM = Limpy Mountain)

```{r}

mod2 <- glm(cbind(nights_STOC_ANY, duration - nights_STOC_ANY) ~ siteID + Distance,
             data = daysMerge, family = quasibinomial(link = 'logit'))
summary(mod2)
# plot(allEffects(mod2))

```

plot:

```{r echo=FALSE}

## plot
plot.mod2 <- ggplot(daysMerge, aes(x = Distance/1000, y = nights_STOC_ANY/duration, 
                                     nights = nights_STOC_4N, total = duration, siteID = siteID)) +
  geom_point(data = daysMerge[daysMerge$nights_STOC_ANY > 0,], #we don't want to plot 0s
             color = 'gray40', size = 2) +
  geom_smooth(method = 'glm', method.args = list(family = 'quasibinomial'),
              formula = cbind(nights, total - nights) ~ x, color = 'aquamarine4', fill = 'aquamarine3') + 
  facet_grid(~nname) +
  ylim(c(0,0.65)) +
    ylab('Proportion of nights with detection') +
    xlab('Distance from center (km)') +
    # ggtitle('Spotted owl detections by reproductive state') +
    theme(panel.background = element_rect(fill = 'transparent'),
          panel.grid = element_line(color = 'gray90'),
          axis.line = element_line(),
          axis.title = element_text(size = 18),
          # axis.title.x = element_blank(),
          strip.text = element_text(size = 16),
          axis.text = element_text(size = 16),
          plot.title = element_text(hjust = 0.5),
          legend.text = element_text(size = 16),
          legend.title = element_blank(),
          legend.position = 'none',
          legend.background = element_rect(fill='transparent'))
  plot.mod2
    #export 1000 x 500

```


### Now do models separately for detections of females / males

#### 3. Quasibinomial with distance from center as covariate (separate for F/M)

```{r}

mod3f <- glm(cbind(nights_STOC_FEMALE, duration-nights_STOC_ANY) ~ Distance,
             data = daysMerge, family = quasibinomial(link = 'logit'))
# summary(mod3f)

mod3m <- glm(cbind(nights_STOC_MALE, duration-nights_STOC_ANY) ~ Distance,
             data = daysMerge, family = quasibinomial(link = 'logit'))
# summary(mod3m)

```

plot:

```{r echo=FALSE}
#make long DF first
daysMergeLong <- melt(daysMerge, id.vars = c('SITE_STN','duration','reproState','Distance','dist_intended',
                                             'siteID','nname'),
                      value.name = 'n')
daysMergeLong$sex <- ifelse(daysMergeLong$variable %in% 'nights_STOC_FEMALE', 'Female',
                            ifelse(daysMergeLong$variable %in% 'nights_STOC_MALE', 'Male', NA))
  
#plot
plot.mod3 <- ggplot(daysMergeLong, aes(x = Distance/1000, y = n/duration,
                                       nights = n, total = duration,
                                       group = sex, color = sex)) +
  geom_point(data = daysMergeLong[daysMergeLong$variable %in% 'nights_STOC_MALE'
                                  & daysMerge$nights_STOC_ANY > 0,], #we don't want to plot 0s
             aes(color = sex), alpha = 0.3, size = 2) +
  geom_point(data = daysMergeLong[daysMergeLong$variable %in% 'nights_STOC_FEMALE'
                                  & daysMerge$nights_STOC_ANY > 0,], #we don't want to plot 0s
             aes(color = sex), alpha = 0.3, size = 2) +
  geom_smooth(data = daysMergeLong[daysMergeLong$variable %in% 'nights_STOC_FEMALE',],
              method = 'glm', method.args = list(family = 'quasibinomial'),
              formula = cbind(nights, total - nights) ~ x, aes(fill = sex)) +  #add fill
  
  geom_smooth(data = daysMergeLong[daysMergeLong$variable %in% 'nights_STOC_MALE',],
            method = 'glm', method.args = list(family = 'quasibinomial'),
              formula = cbind(nights, total - nights) ~ x, aes(fill = sex)) +  #add fill
  scale_color_manual(values = c('tomato','dodgerblue1')) +
  scale_fill_manual(values = c('tomato','dodgerblue1')) +
  # geom_rug(sides = 'b', lwd = 1) +
  # geom_rug(data = daysMergeLong[daysMergeLong$variable %in% 'nights_STOC_MALE' &
  #                                 daysMergeLong$n > 0], color = 'cyan4', sides = "b", outside = TRUE) +
  # geom_rug(data = daysMergeLong[daysMergeLong$variable %in% 'nights_STOC_FEMALE' &
  #                                 daysMergeLong$n > 0], color = 'tomato1', sides = "b") +
  # coord_cartesian(clip = "off") +
  xlab('Distance from center (km)') + ylab('Proportion of nights with detection') + 
  ylim(c(0,0.65)) +
  theme(
    panel.background = element_rect(fill = 'transparent'),
          panel.grid = element_line(color = 'gray90'),
          axis.line = element_line(),
          axis.title = element_text(size = 18),
          # axis.title.x = element_blank(),
          strip.text = element_text(size = 16),
          axis.text = element_text(size = 16),
          plot.title = element_text(hjust = 0.5),
          legend.text = element_text(size = 16),
          legend.title = element_blank(),
          # legend.position = 'none',
          legend.background = element_rect(fill='transparent'))

plot.mod3
# plot(allEffects(mod3))

```


#### 4. Add a fixed effect for AC site (separately for F/M)
##### (CC = Cummins Creek, DC = Drift Creek, BC = Baker Creek, LM = Limpy Mountain)

```{r}

mod4f <- glm(cbind(nights_STOC_FEMALE, duration - nights_STOC_FEMALE) ~ siteID + Distance,
             data = daysMerge, family = quasibinomial(link = 'logit'))
summary(mod4f)
# plot(allEffects(mod4f))

mod4m <- glm(cbind(nights_STOC_MALE, duration - nights_STOC_MALE) ~ siteID + Distance,
             data = daysMerge, family = quasibinomial(link = 'logit'))
summary(mod4m)
# plot(allEffects(mod4m))

```

plot:

```{r echo=FALSE}

plot.mod4 <- ggplot(daysMergeLong, aes(x = Distance/1000, y = n/duration, 
                                       nights = n, total = duration,
                                       group = sex, color = sex)) +
  geom_point(data = daysMergeLong[daysMergeLong$variable %in% 'nights_STOC_MALE'
                                  & daysMerge$nights_STOC_ANY > 0,], #we don't want to plot 0s
             aes(color = sex), alpha = 0.3, size = 2) +
  geom_point(data = daysMergeLong[daysMergeLong$variable %in% 'nights_STOC_FEMALE'
                                  & daysMerge$nights_STOC_ANY > 0,], #we don't want to plot 0s
             aes(color = sex), alpha = 0.3, size = 2) +
  geom_smooth(data = daysMergeLong[daysMergeLong$variable %in% 'nights_STOC_FEMALE',],
              method = 'glm', method.args = list(family = 'quasibinomial'),
              formula = cbind(nights, total - nights) ~ x, aes(fill = sex)) +
  geom_smooth(data = daysMergeLong[daysMergeLong$variable %in% 'nights_STOC_MALE',],
            method = 'glm', method.args = list(family = 'quasibinomial'),
              formula = cbind(nights, total - nights) ~ x, aes(fill = sex)) +
  facet_grid(~nname) +
  scale_color_manual(values = c('tomato','dodgerblue1')) +
  scale_fill_manual(values = c('tomato','dodgerblue1')) +
  # geom_rug(sides = 'b', lwd = 1) +
  # geom_rug(data = daysMergeLong[daysMergeLong$variable %in% 'nights_STOC_MALE' &
  #                                 daysMergeLong$n > 0], color = 'cyan4', sides = "b", outside = TRUE) +
  # geom_rug(data = daysMergeLong[daysMergeLong$variable %in% 'nights_STOC_FEMALE' &
  #                                 daysMergeLong$n > 0], color = 'tomato1', sides = "b") +
  # coord_cartesian(clip = "off") +
  xlab('Distance from center (km)') + ylab('Proportion of nights with detection') + 
  ylim(c(0,0.65)) +
  theme(
    panel.background = element_rect(fill = 'transparent'),
          panel.grid = element_line(color = 'gray90'),
          axis.line = element_line(),
          axis.title = element_text(size = 18),
          # axis.title.x = element_blank(),
          strip.text = element_text(size = 16),
          axis.text = element_text(size = 16),
          plot.title = element_text(hjust = 0.5),
          legend.text = element_text(size = 16),
          legend.title = element_blank(),
          # legend.position = 'none',
          legend.background = element_rect(fill='transparent'))

plot.mod4
  #export 1000 x 500

```


#### 5. ReproState

```{r}

mod5 <- glm(cbind(nights_STOC_ANY, duration - nights_STOC_ANY) ~ reproState,
             data = daysMerge, family = quasibinomial(link = 'logit'))
summary(mod5)
# plot(allEffects(mod5))

mod5f <- glm(cbind(nights_STOC_FEMALE, duration - nights_STOC_FEMALE) ~ reproState,
             data = daysMerge, family = quasibinomial(link = 'logit'))
summary(mod5f)

mod5m <- glm(cbind(nights_STOC_MALE, duration - nights_STOC_MALE) ~ reproState,
             data = daysMerge, family = quasibinomial(link = 'logit'))
summary(mod5m)

```

plot:

```{r echo=FALSE}

## plot

#save effects as dataframe
  df.f.repro <- allEffects(mod5f)
  df.f.repro <- data.frame(df.f.repro$reproState); df.f.repro$sex <- 'female'

  df.m.repro <- allEffects(mod5m)
  df.m.repro <- data.frame(df.m.repro$reproState); df.m.repro$sex <- 'male'

  fm.repro.prop <- rbind(df.f.repro, df.m.repro)
  
  fm.repro.prop$reproState <- factor(fm.repro.prop$reproState,
                                    levels = c('female','pair'))

  #and plot
  plot.mod5.mf <- ggplot(fm.repro.prop, aes(x = reproState, y = fit)) + 
      geom_point(size = 3) +
      geom_errorbar(aes(ymin = lower, ymax = upper)) +
      ylim(c(0,1)) +
      ylab('Probability') +
      facet_grid(~sex) +
      # scale_color_manual(values = c('tomato1', 'steelblue2', 'seagreen')) +
      # ggtitle('Spotted owl detections by reproductive state') +
      theme(panel.background = element_rect(fill = 'transparent'),
            panel.grid = element_line(color = 'gray90'),
            axis.line = element_line(),
            axis.title = element_text(size = 18),
            axis.title.x = element_blank(),
            strip.text = element_text(size = 16),
            axis.text = element_text(size = 16),
            plot.title = element_text(hjust = 0.5),
            legend.text = element_text(size = 16),
            legend.title = element_blank(),
            legend.position = 'none',
            legend.background = element_rect(fill='transparent'))
    plot.mod5.mf


# ALL
#save effects as dataframe
  df.repro <- allEffects(mod5)
  df.repro <- data.frame(df.repro$reproState)
  
  df.repro$reproState <- factor(df.repro$reproState, levels = c('female','pair'))
  
  #and plot
  plot.mod5 <- ggplot(df.repro, aes(x = reproState, y = fit)) + 
      geom_point(size = 3) +
      geom_errorbar(aes(ymin = lower, ymax = upper)) +
      ylim(c(0,1)) +
      ylab('Probability') +
      # scale_color_manual(values = c('tomato1', 'steelblue2', 'seagreen')) +
      # ggtitle('Spotted owl detections by reproductive state') +
      theme(panel.background = element_rect(fill = 'transparent'),
            panel.grid = element_line(color = 'gray90'),
            axis.line = element_line(),
            axis.title = element_text(size = 18),
            axis.title.x = element_blank(),
            strip.text = element_text(size = 16),
            axis.text = element_text(size = 16),
            plot.title = element_text(hjust = 0.5),
            legend.text = element_text(size = 16),
            legend.title = element_blank(),
            legend.position = 'none',
            legend.background = element_rect(fill='transparent'))
    plot.mod5


```



################## 
#### 3. Add a *random* effect for AC site
##### (But can't do 'quasi' family in lmer so this is jut Binomial)
##### -- can adjust later

```{r}

#scale distance variable
daysMerge$distScaled <- scale(daysMerge$Distance)
  # hist(daysMerge$Distance)  
  # hist(daysMerge$distScaled)

mod6 <- glmer(cbind(nights_STOC_ANY, duration - nights_STOC_ANY) ~ (1|siteID) + distScaled,
             data = daysMerge, family = binomial(link = 'logit'))

summary(mod6)
# plot(allEffects(mod6))

```
